Get all the datasets from each of the six celltypes in the ABRIG data.

Get all the datasets from each of the six celltypes in the ABRIG data.

Usage

allrefs(con, pfiles, genelocs, gene)

Arguments

con

A duckdb connection

pfiles

list of absolute paths to the data for each tissue.

genelocs

GenomicRanges instance

gene

A gene from the human genome.

Value

A list of data table by type for the given gene.

Examples

utils::data("gloc_hg19", package = "GGIpack2")
con = DBI::dbConnect(duckdb::duckdb())
options(useFancyQuotes=FALSE)
allrefs( con =con, gene = 'DSP', pfiles =ABRIGparquet_paths(), genelocs = gloc_hg19)
#> Error in ABRIGparquet_paths(): could not find function "ABRIGparquet_paths"
DBI::dbDisconnect(con)