get scores of selected types from a GGIresource instance

get scores of selected types from a GGIresource instance

Usage

topscores_by(res, scope = "chr", val, n = 100)

Arguments

res

GGIresource instance

scope

character(1) one of c("chr", "gene", "rsid")

val

a value in the selected `scope`

n

numeric(1) used to filter results through `head()`

Examples

# we rely on duckdb for query resolution
blex = system.file("parquet", "testbl.parquet", package="GGIpack2")
pp = sprintf("read_parquet(%s)", sQuote(blex, q=FALSE))
con = DBI::dbConnect(duckdb::duckdb())
tb = dplyr::tbl(con, pp)
nres = new("GTExresource", space = "hg19", tbl = tb)
topscores_by(nres, scope="chr", val="1", n=5)
#> Error: Cannot translate a <standardGeneric> object to SQL.
#> ℹ Do you want to force evaluation in R with (e.g.) `!!x` or `local(x)`?
topscores_by(nres, scope="gene", val="FO538757.2", n=5)
#> [1] "FO538757.2"
#> Error: Cannot translate a <standardGeneric> object to SQL.
#> ℹ Do you want to force evaluation in R with (e.g.) `!!x` or `local(x)`?
topscores_by(nres, scope="rsid", val="rs368811019", n=5)
#> Error: Cannot translate a <standardGeneric> object to SQL.
#> ℹ Do you want to force evaluation in R with (e.g.) `!!x` or `local(x)`?
DBI::dbDisconnect(con)